Add User, Team and UserTeam entities and repositories

+++ src/main/java/org/digitalecmt/qualityassurance/repository/TeamRepository.java
@@ -1,48 +1,21 @@
-/**
- * The MIT License (MIT)
- * <p>
- * Copyright (c) 2021 the original author or authors.
- * <p>
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- * <p>
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- * <p>
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
 package org.digitalecmt.qualityassurance.repository;
 
 import java.util.List;
 
 import org.digitalecmt.qualityassurance.dto.TeamWithUsernameDTO;
-import org.digitalecmt.qualityassurance.model.persistence.Team;
+import org.digitalecmt.qualityassurance.models.entities.Team;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 import org.springframework.stereotype.Repository;
 
 /**
- * Repository interface for managing Team entities.
- * <p>
- * Provides methods for team and related data entries.
- * </p>
+ * Repository interface for accessing and managing {@link Team} entities.
+ * Extends {@link JpaRepository} to provide CRUD operations and additional JPA
+ * functionalities.
  */
 @Repository
-public interface TeamRepository
-		extends JpaRepository<Team, Integer> {
-
+public interface TeamRepository extends JpaRepository<Team, Long> {
 	/**
 	 * Retrieves a list of teams with the username of the user who created this
 	 * team.
@@ -74,16 +47,16 @@ public interface TeamRepository
 			"FROM DataEntry d " +
 			"JOIN TeamStudyAccess tsa ON d.studyId = tsa.studyId " +
 			"WHERE tsa.teamId = :teamId")
-	Long countByTeamId(@Param("teamId") Integer teamId);
+	Long countByTeamId(@Param("teamId") Long teamId);
 
 	/**
-     * Retrieves a list of all team IDs.
-     * <p>
-     * Selects the IDs of all teams from the Team entity.
-     * </p>
-     *
-     * @return a list of all team IDs.
-     */
+	 * Retrieves a list of all team IDs.
+	 * <p>
+	 * Selects the IDs of all teams from the Team entity.
+	 * </p>
+	 *
+	 * @return a list of all team IDs.
+	 */
 	@Query("SELECT t.teamName " +
 			"FROM Team t")
 	List<Integer> findTeamIds();
