+++ src/main/java/org/digitalecmt/qualityassurance/repository/UserTeamRepository.java
@@ -27,7 +27,7 @@ package org.digitalecmt.qualityassurance.repository;
 import java.util.List;
 
 import org.digitalecmt.qualityassurance.dto.UserWithTeamDTO;
-import org.digitalecmt.qualityassurance.model.persistence.UserTeam;
+import org.digitalecmt.qualityassurance.models.entities.UserTeam;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -35,20 +35,20 @@ import org.springframework.stereotype.Repository;
 
 @Repository
 public interface UserTeamRepository
-        extends JpaRepository<UserTeam, Integer> {
+		extends JpaRepository<UserTeam, Long> {
 
 	void deleteByUserId(int userId);
-	
-	@Query("SELECT new org.digitalecmt.qualityassurance.dto.UserWithTeamDTO(ut.userId, ut.teamId, t.teamName) " + 
-		       "FROM UserTeam ut " + 	
-		       "JOIN Team t ON ut.teamId = t.teamId")
-		List<UserWithTeamDTO> findUsersWithTeams();
-	
-	@Query("SELECT new org.digitalecmt.qualityassurance.dto.UserWithTeamDTO(ut.userId, ut.teamId, t.teamName) " + 
-	           "FROM UserTeam ut " + 	
-	           "JOIN Team t ON ut.teamId = t.teamId " +
-	           "WHERE ut.userId = :userId")
-	    List<UserWithTeamDTO> findUserTeamsByUserId(@Param("userId") Integer userId);
+
+	@Query("SELECT new org.digitalecmt.qualityassurance.dto.UserWithTeamDTO(ut.userId, ut.teamId, t.teamName) " +
+			"FROM UserTeam ut " +
+			"JOIN Team t ON ut.teamId = t.teamId")
+	List<UserWithTeamDTO> findUsersWithTeams();
+
+	@Query("SELECT new org.digitalecmt.qualityassurance.dto.UserWithTeamDTO(ut.userId, ut.teamId, t.teamName) " +
+			"FROM UserTeam ut " +
+			"JOIN Team t ON ut.teamId = t.teamId " +
+			"WHERE ut.userId = :userId")
+	List<UserWithTeamDTO> findUserTeamsByUserId(@Param("userId") Long userId);
 
 	List<UserTeam> findByUserId(int userId);
 }
